pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "panparsex"
version = "0.1.0"
description = "Universal, extensible parser for many file types and websites with a clean plugin architecture."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [{name="Dhruvil Darji", email="dhruvildarji1409@gmail.com"}]
keywords = ["parser", "ingestion", "etl", "unstructured", "scraper", "crawler", "pdf", "html", "json", "yaml", "xml", "text"]
dependencies = [
  "pydantic>=2.5",
  "beautifulsoup4>=4.12",
  "lxml>=5.0",
  "html5lib>=1.1",
  "requests>=2.31",
  "tqdm>=4.66",
  "markdown-it-py>=3.0",
  "pypdf>=3.0",
  "pdfminer.six>=20221105",
  "PyYAML>=6.0",
  "python-docx>=0.8.11",
  "openpyxl>=3.1.0",
  "python-pptx>=0.6.21",
  "python-magic>=0.4.27",
  "chardet>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/dhruvildarji/panparsex"
Repository = "https://github.com/dhruvildarji/panparsex"
Issues = "https://github.com/dhruvildarji/panparsex/issues"
Documentation = "https://github.com/dhruvildarji/panparsex#readme"

[project.scripts]
panparsex = "panparsex.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "safety>=2.0",
    "bandit>=1.7",
]

[project.entry-points."panparsex.parsers"]
# Third-party plugins can register new parsers here (module:function returning a Parser instance)
# e.g., "docx = 'my_pkg.my_docx:make_parser'"
